Exercise 12.10
setAsync function takes in key and value, using the key to store the
value.
getAsync function takes in a key and returns the value in a promise.
Implement a todo counter that saves the number of created todos to
Redis:

Step 1: Whenever a request is sent to add a todo, increment the counter
by one.
Step 2: Create a GET /statistics endpoint where you can ask for the
usage metadata. The format should be the following JSON:
{
  "added_todos": 0
}